test-requirements.txt
hacking>=0.12.0,<0.13 # Apache-2.0

betamax
coverage>=4.0 # Apache-2.0
fixtures>=3.0.0
mock
python-subunit>=0.0.18 # Apache-2.0/BSD
requests-mock[fixture]
sphinx>=1.2.1,!=1.3b1,<1.4 # BSD
testrepository>=0.0.18  # Apache-2.0/BSD
testscenarios>=0.4  # Apache-2.0/BSD
testtools>=1.4.0 # MIT
PyYAML
